requirements.txt
flask==2.3.2
flask-cors==3.0.10
waitress==2.1.2
watchdog==2.1.6
python-dotenv==0.21.0
#torch==1.13.1
#transformers==4.26.1
numpy==1.24.2